Relative path fix
Fix relative paths specified in the config file not being converted into absolute paths using the correct root in some Linux systems.
Also fixed the template directory not being a sub-directory of the data directory. I may be wrong on this one, but the directory structure and the example conf.ini leaded me to this conclusion.